Add fullscale and maximum normalization tiffio now includes matrix Change tiff_frommatrix function to discretize the double matrix

+++ release/src/gerchberg.c
@@ -99,8 +99,8 @@ double identity(double d, void **args) {return d;}
  *
  */
 double div_dim(double d, void **args) {
-  double dim = (double) *((int*) args[0]);
-  return d/(dim*dim);
+  int dim = *((int*) args[0]);
+  return d/(double) (dim*dim);
 }
 
 /**
